refactor(entry): mutation hookを責務別に分割 (#1306)
